Rename Polyclinic.Domain
