+++ Polyclinic/Polyclinic.Api/Polyclinic.Api.csproj
@@ -14,7 +14,6 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\Polyclinic.Contracts\Polyclinic.Contracts.csproj" />
-		<ProjectReference Include="..\Polyclinic\Polyclinic.csproj" />
 		<ProjectReference Include="..\Polyclinic.Application\Polyclinic.Application.csproj" />
 		<ProjectReference Include="..\Polyclinic.Repositories\Polyclinic.Repositories.csproj" />
 	</ItemGroup>
