+++ Polyclinic/Polyclinic.Application/Service/AppointmentService.cs
@@ -2,7 +2,7 @@
 using Polyclinic.Contracts.Appointments;
 using Polyclinic.Contracts.Patients;
 using Polyclinic.Contracts.Doctors;
-using Polyclinic.Models;
+using Polyclinic.Domain.Models;
 using Polyclinic.Repositories;
 
 namespace Polyclinic.Application.Service;
