+++ Polyclinic/Polyclinic.Application/Service/DoctorService.cs
@@ -2,7 +2,7 @@
 using Polyclinic.Contracts.Doctors;
 using Polyclinic.Contracts.Appointments;
 using Polyclinic.Contracts.Patients;
-using Polyclinic.Models;
+using Polyclinic.Domain.Models;
 using Polyclinic.Repositories;
 
 namespace Polyclinic.Application.Service;
