+++ Polyclinic/Polyclinic.Application/Service/SpecializationService.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using Polyclinic.Contracts.Specializations;
-using Polyclinic.Models;
+using Polyclinic.Domain.Models;
 using Polyclinic.Repositories;
 
 namespace Polyclinic.Application.Service;
