+++ Polyclinic/Polyclinic.Domain/Polyclinic.Domain.csproj
@@ -6,5 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-
 </Project>
