+++ Polyclinic/Polyclinic.Repositories/Polyclinic.Repositories.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\Polyclinic\Polyclinic.csproj" />
+		<ProjectReference Include="..\Polyclinic.Domain\Polyclinic.Domain.csproj" />
 		<ProjectReference Include="..\Polyclinic.Test\Polyclinic.Test.csproj" />
 	</ItemGroup>
