+++ Polyclinic/Polyclinic.Test/Polyclinic.Test.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -21,8 +21,8 @@
     <Using Include="Xunit" />
   </ItemGroup>
 
-  <ItemGroup>
-	<ProjectReference Include="..\Polyclinic\Polyclinic.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\Polyclinic.Domain\Polyclinic.Domain.csproj" />
+	</ItemGroup>
 
 </Project>
