+++ Polyclinic/Polyclinic.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.14.36511.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic", "Polyclinic\Polyclinic.csproj", "{5566D06A-6A2B-4193-83E5-F82A724F9959}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic.Test", "Polyclinic.Test\Polyclinic.Test.csproj", "{8CF65AE6-98F7-470B-B41B-CDA372BE3E1A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic.Contracts", "Polyclinic.Contracts\Polyclinic.Contracts.csproj", "{BB55172C-8968-4F07-A77E-4E5F6C1384EE}"
@@ -15,16 +13,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic.Api", "Polyclini
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic.Repositories", "Polyclinic.Repositories\Polyclinic.Repositories.csproj", "{2CE77193-C1D9-42F6-991B-4CE7D07EBC80}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Polyclinic.Domain", "Polyclinic.Domain\Polyclinic.Domain.csproj", "{B5A60F34-DAB5-48DF-9CFF-DA98921FD6D6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5566D06A-6A2B-4193-83E5-F82A724F9959}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5566D06A-6A2B-4193-83E5-F82A724F9959}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5566D06A-6A2B-4193-83E5-F82A724F9959}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5566D06A-6A2B-4193-83E5-F82A724F9959}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8CF65AE6-98F7-470B-B41B-CDA372BE3E1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8CF65AE6-98F7-470B-B41B-CDA372BE3E1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8CF65AE6-98F7-470B-B41B-CDA372BE3E1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -45,6 +41,10 @@ Global
 		{2CE77193-C1D9-42F6-991B-4CE7D07EBC80}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2CE77193-C1D9-42F6-991B-4CE7D07EBC80}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2CE77193-C1D9-42F6-991B-4CE7D07EBC80}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5A60F34-DAB5-48DF-9CFF-DA98921FD6D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5A60F34-DAB5-48DF-9CFF-DA98921FD6D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5A60F34-DAB5-48DF-9CFF-DA98921FD6D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5A60F34-DAB5-48DF-9CFF-DA98921FD6D6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
